seperated convertorsToGraph
